inference-api-example/requirements.txt
fastapi==0.75.1
uvicorn==0.17.6
amqpstorm==2.10.4
python-dotenv==0.20.0